# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NetflixAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Classification": {
      "get": {
        "tags": [
          "Classification"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Classification"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Classification"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Classification"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Classification"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Classification"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Classification"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              }
            }
          }
        }
      }
    },
    "/api/Classification/{id}": {
      "get": {
        "tags": [
          "Classification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Classification"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Classification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Classification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Classification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Country": {
      "get": {
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Country"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          }
        }
      }
    },
    "/api/Country/{id}": {
      "get": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Country"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Classification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ageRating": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}